LayoutTests/webrtc/video-replace-track.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <title>Testing basic video exchange from offerer to receiver</title>
        <script src="../resources/testharness.js"></script>
        <script src="../resources/testharnessreport.js"></script>
    </head>
    <body>
<div id="log"></div>
        <video id="video" autoplay=""></video>
        <canvas id="canvas" width="640" height="480"></canvas>
        <script src ="routines.js"></script>
        <script>
video = document.getElementById("video");
canvas = document.getElementById("canvas");

function grabImagePixels()
{
    canvas.width = video.videoWidth;
    canvas.height = video.videoHeight;
    canvas.getContext('2d').drawImage(video, 0, 0, canvas.width, canvas.height);

    imageData = canvas.getContext('2d').getImageData(20, 20, 2, 2);
    return imageData.data;
 }

function testFrontCameraImage()
{
    data = grabImagePixels();

    assert_true(data[0] < 20);
    assert_true(data[1] < 20);
    assert_true(data[2] < 20);
}

function testBackCameraImage()
{
    data = grabImagePixels();

    assert_true(data[0] > 100);
    assert_true(data[1] > 100);
    assert_true(data[2] > 100);

    assert_true(data[0] < 200);
    assert_true(data[1] < 200);
    assert_true(data[2] < 200);
}

promise_test((test) => {
    if (window.testRunner)
        testRunner.setUserMediaPermission(true);

    var sender;
    var frontStream;
    var backStream;
    return navigator.mediaDevices.getUserMedia({ video: { facingMode: { exact: ["user"] } } }).then((stream) => {
        frontStream = stream;
        return navigator.mediaDevices.getUserMedia({ video: { facingMode: { exact: ["environment"] } } });
    }).then((stream) => {
        backStream = stream;
    }).then(() => {
        return new Promise((resolve, reject) => {
            if (window.internals)
                internals.useMockRTCPeerConnectionFactory("TwoRealPeerConnections");

            createConnections((firstConnection) => {
                sender = firstConnection.addTrack(frontStream.getVideoTracks()[0], frontStream);
            }, (secondConnection) => {
                secondConnection.ontrack = (trackEvent) => {
                    resolve(trackEvent.streams[0]);
                };
            });
            setTimeout(() => reject("Test timed out"), 5000);
        });
    }).then((remoteStream) => {
        video.srcObject = remoteStream;
        return video.play();
    }).then(() => {
        testFrontCameraImage();
    }).then(() => {
        var currentTrack = sender.track;
        promise = sender.replaceTrack(backStream.getVideoTracks()[0]);
        assert_true(currentTrack === sender.track);
        return promise;
    }).then(() => {
        assert_true(sender.track === backStream.getVideoTracks()[0]);
        return waitFor(500);
    }).then(() => {
        testBackCameraImage();
    });
}, "Switching from front to back camera");

promise_test((test) => {
    if (window.testRunner)
        testRunner.setUserMediaPermission(true);

    var sender;
    var frontStream;
    var backStream;

    return navigator.mediaDevices.getUserMedia({ video: { height: { min: 400 }, facingMode: { exact: ["user"] } } }).then((stream) => {
        frontStream = stream;
        return navigator.mediaDevices.getUserMedia({ video: { height: { max: 400 }, facingMode: { exact: ["environment"] } } });
    }).then((stream) => {
        backStream = stream;
        assert_true(frontStream.getVideoTracks()[0].getSettings().height >= 400, "frontStream should be bigger");
        assert_true(backStream.getVideoTracks()[0].getSettings().height < 400, "backStream should be smaller");
    }).then(() => {
        return new Promise((resolve, reject) => {
            if (window.internals)
                internals.useMockRTCPeerConnectionFactory("TwoRealPeerConnections");

            createConnections((firstConnection) => {
                sender = firstConnection.addTrack(frontStream.getVideoTracks()[0], frontStream);
            }, (secondConnection) => {
                secondConnection.ontrack = (trackEvent) => {
                    resolve(trackEvent.streams[0]);
                };
            });
            setTimeout(() => reject("Test timed out"), 5000);
        });
    }).then((remoteStream) => {
        video.srcObject = remoteStream;
        return video.play();
    }).then(() => {
        testFrontCameraImage();
    }).then(() => {
        return sender.replaceTrack(backStream.getVideoTracks()[0]);
    }).then(() => {
        return waitFor(500);
    }).then(() => {
        testBackCameraImage();
    });
}, "Switching from front to back camera, with lower resolution");

promise_test((test) => {
    if (window.testRunner)
        testRunner.setUserMediaPermission(true);

    var sender;
    var frontStream;
    var backStream;

    return navigator.mediaDevices.getUserMedia({ video: { height: { max: 400 }, facingMode: { exact: ["user"] } } }).then((stream) => {
        frontStream = stream;
        return navigator.mediaDevices.getUserMedia({ video: { height: { min: 400 }, facingMode: { exact: ["environment"] } } });
    }).then((stream) => {
        backStream = stream;
        assert_true(frontStream.getVideoTracks()[0].getSettings().height < 400, "front stream should be smaller");
        assert_true(backStream.getVideoTracks()[0].getSettings().height >= 400, "back stream should be bigger");
    }).then(() => {
        return new Promise((resolve, reject) => {
            if (window.internals)
                internals.useMockRTCPeerConnectionFactory("TwoRealPeerConnections");

            createConnections((firstConnection) => {
                sender = firstConnection.addTrack(frontStream.getVideoTracks()[0], frontStream);
            }, (secondConnection) => {
                secondConnection.ontrack = (trackEvent) => {
                    resolve(trackEvent.streams[0]);
                };
            });
            setTimeout(() => reject("Test timed out"), 5000);
        });
    }).then((remoteStream) => {
        video.srcObject = remoteStream;
        return video.play();
    }).then(() => {
        testFrontCameraImage();
    }).then(() => {
        return promise_rejects(test, "InvalidModificationError", sender.replaceTrack(backStream.getVideoTracks()[0]));
    }).then(() => {
        return waitFor(500);
    }).then(() => {
        testBackCameraImage();
    });
}, "Switching from front to back camera, with higher resolution");
        </script>
    </body>
</html>
